+++ packages/astro/src/actions/consts.ts
@@ -1,0 +1,3 @@
+export const VIRTUAL_MODULE_ID = 'astro:actions';
+export const RESOLVED_VIRTUAL_MODULE_ID = '\0' + VIRTUAL_MODULE_ID;
+export const ACTIONS_TYPES_FILE = 'actions.d.ts';